requirements/base.txt
django
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
Pillow
